--- a/changes/6987.fixed
+++ b/changes/6987.fixed
@@ -1,0 +1,1 @@
+Improved speed of IPAM 1.x-to-2.x data migration `0031_ipam___data_migrations` by about 20% when working with large data sets.