--- a/changes/6987.housekeeping
+++ b/changes/6987.housekeeping
@@ -1,0 +1,7 @@
+Added `django-test-migrations` as a testing dependency.
+Added `nautobot.ipam.tests.test_migrations.IPAMDataMigration0031TestCase` that can be run to reproducibly test IPAM data migration `0031_ipam___data_migrations` for correctness and performance.
+Renamed `invoke tests` task to `invoke lint` and renamed `invoke unittest` task to `invoke tests`.
+Changed 'invoke tests' task to default to _not_ measuring and reporting code coverage by default, since doing so slows test performance.
+Removed `django-slowtests` as a development dependency and removed associated test-performance-measuring functionality.
+Removed distinct `invoke integration-test`, `invoke performance-test`, and `invoke unittest-coverage` tasks; integration tests should now be run with `invoke tests --tag integration` and test coverage is reported when running `invoke tests --coverage`.
+Refactored `nautobot.ipam.tests.test_migrations.AggregateToPrefixMigrationTestCase` to use `django-test-migrations` and be runnable.