--- a/nautobot/ipam/tests/test_migrations.py
+++ b/nautobot/ipam/tests/test_migrations.py
@@ -488,7 +488,7 @@ class AggregateToPrefixMigrationTestCase(MigratorTestCase):
 
 class IPAMDataMigration0031TestCase(MigratorTestCase):
     migrate_from = ("ipam", "0030_ipam__namespaces")
-    migrate_to = ("ipam", "0031_ipam___data_migrations")
+    migrate_to = ("ipam", "0032_ipam__namespaces_finish")
 
     def prepare(self):
         # Create an arbitrary set of prefixes and IPs mostly subdividing and consuming the given subnet, including dupes