# Closes #6987 
# What's Changed

- Improved speed of IPAM 1.x-to-2.x data migration `0031_ipam___data_migrations` by about 20% when working with large data sets.
   - Note the bulk of this migration is actually defined in `nautobot/ipam/utils/migration.py`, which is the file with the most important changes for purposes of this PR.
- Added `django-test-migrations` as a testing dependency.
- Added `nautobot.ipam.tests.test_migrations.IPAMDataMigration0031TestCase` that can be run to reproducibly test IPAM data migration `0031_ipam___data_migrations` for correctness and performance.
- Renamed `invoke tests` task to `invoke lint` and renamed `invoke unittest` task to `invoke tests`.
   - This aligns us a bit more closely with the current internal ADR for development environments
- Changed `invoke tests` task to default to _not_ measuring and reporting code coverage by default, since doing so slows test performance.
   - This aligns with a similar recent change in the App dev environments and CI.
- Removed `django-slowtests` as a development dependency and removed associated test-performance-measuring functionality.
   - While a nice idea, this code hasn't been maintained or really used since it was added in 2022, and we have a different direction planned for application-level performance tests at this point.
- Removed distinct `invoke integration-test`, `invoke performance-test`, and `invoke unittest-coverage` tasks; integration tests should now be run with `invoke tests --tag integration` and test coverage is reported when running `invoke tests --coverage`.
   - This aligns us a bit more closely with the current internal ADR for development environments, and eliminates error-prone duplication of code and parameters across these distinct task definitions.
- Refactored `nautobot.ipam.tests.test_migrations.AggregateToPrefixMigrationTestCase` to use `django-test-migrations` and be runnable.

# Screenshots

## Before optimizations:

```
>>> Checking whether any Interface or VMInterface has IPs with differing VRFs...
    ... completed (elapsed time: 0.0 seconds)
>>> Verifying all Prefix.broadcast values...
    ... completed (elapsed time: 0.6 seconds)
>>> Setting Prefix.version and IPAddress.version values...
    ... completed (elapsed time: 0.5 seconds)
>>> Processing VRFs...
    VRF '10.0.0.0/14 VRF 6' migrated from Namespace 'Global' to 'Cleanup Namespace (1)'
    VRF '10.0.0.0/14 VRF 8' migrated from Namespace 'Global' to 'Cleanup Namespace (2)'
    VRF '10.0.0.0/14 VRF 4' migrated from Namespace 'Global' to 'Cleanup Namespace (1)'
    VRF '10.0.0.0/14 VRF 2' migrated from Namespace 'Global' to 'Cleanup Namespace (3)'
    VRF '10.0.0.0/14 VRF 1' migrated from Namespace 'Global' to 'Cleanup Namespace (2)'
    VRF '10.0.0.0/14 VRF 5' migrated from Namespace 'Global' to 'Cleanup Namespace (4)'
    VRF '10.0.0.0/14 VRF 10' migrated from Namespace 'Global' to 'Cleanup Namespace (3)'
    VRF '10.0.0.0/14 VRF 9' migrated from Namespace 'Global' to 'Cleanup Namespace (1)'
    VRF '10.0.0.0/14 VRF 3' migrated from Namespace 'Global' to 'Cleanup Namespace (5)'
    VRF '10.0.0.0/14 VRF 7' migrated from Namespace 'Global' to 'Cleanup Namespace (3)'
    ... completed (elapsed time: 0.8 seconds)
>>> Processing IPAddresses...
    >>> Reparenting individual IPAddresses to a close-enough parent Prefix...
        ... completed (elapsed time: 218.5 seconds)  <<<<<<<<<<<<<<<<<<<<<<<<<<<<<
    >>> Reparenting orphaned IPAddresses by creating new Prefixes as needed...
        ... completed (elapsed time: 9.4 seconds)
    ... completed (elapsed time: 227.9 seconds)
>>> Processing duplicate Prefixes...
    ... completed (elapsed time: 6.8 seconds)
>>> Reparenting Prefixes...
    ... completed (elapsed time: 53.3 seconds)  <<<<<<<<<<<<<<<<<<<<<<<<<<<<<
>>> Copying VRFs to cleanup Namespaces as needed...
    ... completed (elapsed time: 0.3 seconds)
>>> Processing Interfaces...
    ... completed (elapsed time: 0.0 seconds)
>>> Processing VRF to Prefix many-to-many...
    ... completed (elapsed time: 0.7 seconds)
```

## After optimizations:

```
>>> Checking whether any Interface or VMInterface has IPs with differing VRFs...
    ... completed (elapsed time: 0.0 seconds)
>>> Verifying all Prefix.broadcast values...
    ... completed (elapsed time: 0.4 seconds)
>>> Setting Prefix.version and IPAddress.version values...
    ... completed (elapsed time: 0.5 seconds)
>>> Processing VRFs...
    VRF '10.0.0.0/14 VRF 6' migrated from Namespace 'Global' to 'Cleanup Namespace (1)'
    VRF '10.0.0.0/14 VRF 8' migrated from Namespace 'Global' to 'Cleanup Namespace (2)'
    VRF '10.0.0.0/14 VRF 4' migrated from Namespace 'Global' to 'Cleanup Namespace (1)'
    VRF '10.0.0.0/14 VRF 2' migrated from Namespace 'Global' to 'Cleanup Namespace (3)'
    VRF '10.0.0.0/14 VRF 1' migrated from Namespace 'Global' to 'Cleanup Namespace (2)'
    VRF '10.0.0.0/14 VRF 5' migrated from Namespace 'Global' to 'Cleanup Namespace (4)'
    VRF '10.0.0.0/14 VRF 10' migrated from Namespace 'Global' to 'Cleanup Namespace (3)'
    VRF '10.0.0.0/14 VRF 3' migrated from Namespace 'Global' to 'Cleanup Namespace (5)'
    VRF '10.0.0.0/14 VRF 9' migrated from Namespace 'Global' to 'Cleanup Namespace (1)'
    VRF '10.0.0.0/14 VRF 7' migrated from Namespace 'Global' to 'Cleanup Namespace (3)'
    ... completed (elapsed time: 0.9 seconds)
>>> Processing IPAddresses...
    >>> Reparenting individual IPAddresses to a close-enough parent Prefix...
        ... completed (elapsed time: 167.3 seconds)  <<<<<<<<<<<<<<<<<<<<<<<<<<<<<
    >>> Reparenting orphaned IPAddresses by creating new Prefixes as needed...
        ... completed (elapsed time: 10.5 seconds)
    ... completed (elapsed time: 177.8 seconds)
>>> Processing duplicate Prefixes...
    ... completed (elapsed time: 6.7 seconds)
>>> Reparenting Prefixes...
    ... completed (elapsed time: 42.2 seconds)  <<<<<<<<<<<<<<<<<<<<<<<<<<<<<
>>> Copying VRFs to cleanup Namespaces as needed...
    ... completed (elapsed time: 0.3 seconds)
>>> Processing Interfaces...
    ... completed (elapsed time: 0.0 seconds)
>>> Processing VRF to Prefix many-to-many...
    ... completed (elapsed time: 0.7 seconds)
```

# TODO
<!--
    Please feel free to update todos to keep track of your own notes for WIP PRs.
-->
- [x] Explanation of Change(s)
- [x] Added change log fragment(s) (for more information see [the documentation](https://docs.nautobot.com/projects/core/en/stable/development/core/#creating-changelog-fragments))
- [x] Attached Screenshots, Payload Example
- [x] Unit, Integration Tests
- [x] Documentation Updates (when adding/changing features)
- n/a Example App Updates (when adding/changing features)
- [x] Outline Remaining Work, Constraints from Design
- [ ] (MAYBE) add `invoke tests --tag migration_test` to CI? (Need to discuss this with the team)
- [ ] Enhance documentation about running migration tests?
- [ ] Rename `invoke migration-test` task to something a bit clearer as to its (narrow) purpose?
- [x] Self-review the diffs below with a fresh eye next week
- [x] Review CI results for unit/integration tests for correctness and compare to `develop` baseline.
- [x] Compare migration performance under MySQL (the above data is from Postgres).