@glennmatthews 2025-03-17
MySQL data points:

## Before

```
>>> Checking whether any Interface or VMInterface has IPs with differing VRFs...
    ... completed (elapsed time: 0.0 seconds)
>>> Verifying all Prefix.broadcast values...
    ... completed (elapsed time: 0.4 seconds)
>>> Setting all Prefix.version and IPAddress.version values...
    ... completed (elapsed time: 25.0 seconds)
>>> Processing VRFs...
    VRF '10.0.0.0/14 VRF 6' migrated from Namespace 'Global' to 'Cleanup Namespace (1)'
    VRF '10.0.0.0/14 VRF 8' migrated from Namespace 'Global' to 'Cleanup Namespace (2)'
    VRF '10.0.0.0/14 VRF 4' migrated from Namespace 'Global' to 'Cleanup Namespace (1)'
    VRF '10.0.0.0/14 VRF 1' migrated from Namespace 'Global' to 'Cleanup Namespace (2)'
    VRF '10.0.0.0/14 VRF 2' migrated from Namespace 'Global' to 'Cleanup Namespace (3)'
    VRF '10.0.0.0/14 VRF 5' migrated from Namespace 'Global' to 'Cleanup Namespace (4)'
    VRF '10.0.0.0/14 VRF 10' migrated from Namespace 'Global' to 'Cleanup Namespace (3)'
    VRF '10.0.0.0/14 VRF 3' migrated from Namespace 'Global' to 'Cleanup Namespace (5)'
    VRF '10.0.0.0/14 VRF 9' migrated from Namespace 'Global' to 'Cleanup Namespace (1)'
    VRF '10.0.0.0/14 VRF 7' migrated from Namespace 'Global' to 'Cleanup Namespace (3)'
    ... completed (elapsed time: 1.5 seconds)
>>> Processing IPAddresses...
    >>> Reparenting individual IPAddresses to a close-enough parent Prefix...
        ... completed (elapsed time: 429.0 seconds)
    >>> Reparenting orphaned IPAddresses by creating new Prefixes as needed...
        ... completed (elapsed time: 9.0 seconds)
    ... completed (elapsed time: 438.0 seconds)
>>> Processing duplicate Prefixes...
    ... completed (elapsed time: 14.5 seconds)
>>> Reparenting Prefixes...
    ... completed (elapsed time: 73.2 seconds)
>>> Copying VRFs to cleanup Namespaces as needed...
    ... completed (elapsed time: 0.2 seconds)
>>> Processing Interfaces...
    ... completed (elapsed time: 0.0 seconds)
>>> Processing VRF to Prefix many-to-many...
    ... completed (elapsed time: 0.4 seconds)
```

## After

```
>>> Checking whether any Interface or VMInterface has IPs with differing VRFs...
    ... completed (elapsed time: 0.0 seconds)
>>> Verifying all Prefix.broadcast values...
    ... completed (elapsed time: 0.4 seconds)
>>> Setting Prefix.version and IPAddress.version values...
    ... completed (elapsed time: 0.8 seconds)
>>> Processing VRFs...
    VRF '10.0.0.0/14 VRF 6' migrated from Namespace 'Global' to 'Cleanup Namespace (1)'
    VRF '10.0.0.0/14 VRF 8' migrated from Namespace 'Global' to 'Cleanup Namespace (2)'
    VRF '10.0.0.0/14 VRF 4' migrated from Namespace 'Global' to 'Cleanup Namespace (1)'
    VRF '10.0.0.0/14 VRF 1' migrated from Namespace 'Global' to 'Cleanup Namespace (2)'
    VRF '10.0.0.0/14 VRF 2' migrated from Namespace 'Global' to 'Cleanup Namespace (3)'
    VRF '10.0.0.0/14 VRF 5' migrated from Namespace 'Global' to 'Cleanup Namespace (4)'
    VRF '10.0.0.0/14 VRF 10' migrated from Namespace 'Global' to 'Cleanup Namespace (3)'
    VRF '10.0.0.0/14 VRF 3' migrated from Namespace 'Global' to 'Cleanup Namespace (5)'
    VRF '10.0.0.0/14 VRF 9' migrated from Namespace 'Global' to 'Cleanup Namespace (1)'
    VRF '10.0.0.0/14 VRF 7' migrated from Namespace 'Global' to 'Cleanup Namespace (3)'
    ... completed (elapsed time: 1.0 seconds)
>>> Processing IPAddresses...
    >>> Reparenting individual IPAddresses to a close-enough parent Prefix...
        ... completed (elapsed time: 296.2 seconds)
    >>> Reparenting orphaned IPAddresses by creating new Prefixes as needed...
        ... completed (elapsed time: 11.8 seconds)
    ... completed (elapsed time: 308.0 seconds)
>>> Processing duplicate Prefixes...
    ... completed (elapsed time: 14.5 seconds)
>>> Reparenting Prefixes...
    ... completed (elapsed time: 97.1 seconds)
>>> Copying VRFs to cleanup Namespaces as needed...
    ... completed (elapsed time: 0.2 seconds)
>>> Processing Interfaces...
    ... completed (elapsed time: 0.0 seconds)
>>> Processing VRF to Prefix many-to-many...
    ... completed (elapsed time: 0.4 seconds)
```